[Thesis] thesis handler
