--- views/profile/thesisinfo.hbs
@@ -1,64 +1,0 @@
-<!-- {{ user.local.title }}
-{{	user }}
-{{	user.education 	}} -->
- <div class="col-md-8">
- 				
-                <h3>Thesis Information</h3>
-                <form class="form-horizontal" role="form">
-                    </div>
-
-                    <div id="tab_2"><table class="table muffin table-striped table-hover margin-bottom0">
-                    <thead>
-                        <tr>
-                        <th>Index</th>
-                        <th><a href="javascript:table_sort('tab_2',1)">Name</a></th>
-                        <th><a href="javascript:table_sort('tab_2',2)">Academic Year</a></th>
-                        <th><a href="javascript:table_sort('tab_2',3)">Advisor</a></th>
-                        <th><a href="javascript:table_sort('tab_2',4)">Co-advisor</a></th>
-                        <th><a href="javascript:table_sort('tab_2',6)">Delete</a></th>
-                        </tr>
-                    </thead>
-                    <tbody>
-                    {{#each user.thesis}}
-                        <tr>                       
-                        <td>{{inc @index}}</td>
-						<td>{{name}}</td>
-						<td>{{academic_year}}</td>
-						<td>{{Advisors}}</td>
-						<td>{{Co-Advisors}}</td>
-                        <td><a href='/deledu?id={{id}}&user={{user.local.username}}'>delete</a></td>
-                        </tr>
-                        <tr>
-                    {{/each}}  
-                    </tbody>
-                    </table>
-                    </div>
-                    
-                    
-                    <!-- end: table full -->
-                    <div class="table-foot muffin"> 
-                        
-                        <div class="pull-right">
-                        <div class="btn-group">
-                            <button id="tab_2_unselect" class="btn btn-mini" type="button"><i class="glyphicon glyphicon-remove"></i></button>
-                            <button id="tab_2_selectall" class="btn btn-mini" type="button">All</button>
-                        </div>
-                        
-                        <div class="btn-group">
-                            <a href = "/addthesis?user={{user.local.username}}"><button id="tab_2_add" class="btn btn-mini" type="button">Add</button></a>
-                        </div>
-                        
-                                    
-                        </div>
-                        
-                        
-                        
-                    </div>
-
-                </form>
-                
-
-        </div>
-
-
-
