+++ views/profile/works/editthesis.hbs
@@ -87,7 +87,7 @@
 	
 	
 </script>
-<form action="/addthesis" method="post" id = "myForm" >
+<form action="/thesisinf/addthesis" method="post" id = "myForm" >
 
 	<div class="row content"> 
                 <form class="form-horizontal" role="form">
